adding gradient colors. Note that interpolator class is not useful for now, but I decided to keep it in case we can need it in the future for grid oversamplin

+++ widget.h
@@ -19,7 +19,7 @@ public:
 
 protected:
     void paintEvent(QPaintEvent *event);
-    void paintRasterOnQimage(QPainter *painter, int zoom=1);
+    void paintRasterOnQimage(QPainter *painter, double zoom=1.0);
 
 private:
     void drawMyGrid();
